Round out the line directive support (#53)
This rounds out a few of the corner cases in `#line` directive support. 

The one issue that remains is when the user combines `/embed` and `#line` directives with full paths the resulting compilation is not 100% portable between machines. There is simply no way to redirect where the compiler attempts to read the target off of disk hence we're stuck with the original path. Filed [a bug](https://github.com/dotnet/roslyn/issues/69659) with roslyn to track fixing this.

+++ src/Basic.CompilerLog.UnitTests/TempDir.cs
@@ -23,7 +23,24 @@ internal sealed class TempDir : IDisposable
 
     public void Dispose()
     {
-        Directory.Delete(DirectoryPath, recursive: true);
+        if (Directory.Exists(DirectoryPath))
+        {
+            Directory.Delete(DirectoryPath, recursive: true);
+        }
+    }
+
+    public string NewFile(string fileName, string content)
+    {
+        var filePath = Path.Combine(DirectoryPath, fileName);
+        File.WriteAllText(filePath, content);
+        return filePath;
+    }
+
+    public string NewDirectory(string name)
+    {
+        var path = Path.Combine(DirectoryPath, name);
+        _ = Directory.CreateDirectory(path);
+        return path;
     }
 
     public void EmptyDirectory()
